test(treedb): make snapshot rebind coverage platform-aware
